refactor(routes): align file/component names with page identities
- Pricing → Models (components/models/Pricing.tsx → Models.tsx)
- UpdatesPage → NewsFaq (layout/updates-page.tsx → news-faq/news-faq.tsx)
- components/api-keys/ → components/keys/

All hash route aliases preserved in pageFromHash (#pricing, #news, #api-keys-* etc still resolve). Server /api/api-keys endpoint untouched.
