+++ enter.pollinations.ai/src/client/components/activity/earnings-visibility.ts
@@ -1,4 +1,4 @@
-import type { ApiKey } from "../api-keys/types.ts";
+import type { ApiKey } from "../keys/types.ts";
 
 export type EarningsApp = {
     id: string;
