+++ enter.pollinations.ai/src/client/components/internal/design-showcase.tsx
@@ -3,9 +3,9 @@ import { cn } from "@/util.ts";
 import { Chart } from "../activity/chart.tsx";
 import { MultiSelect } from "../activity/multi-select.tsx";
 import type { DataPoint } from "../activity/types.ts";
-import { ModalityButton } from "../api-keys/modality-button.tsx";
-import { MODALITY_COLORS, type Modality } from "../api-keys/modality-ui.ts";
 import { Button } from "../button.tsx";
+import { ModalityButton } from "../keys/modality-button.tsx";
+import { MODALITY_COLORS, type Modality } from "../keys/modality-ui.ts";
 import { type ThemeName, themes } from "../layout/dashboard-theme.ts";
 import { Chip } from "../ui/chip.tsx";
 import { IconButton } from "../ui/icon-button.tsx";
