+++ enter.pollinations.ai/src/client/components/models/index.ts
@@ -1,2 +1,2 @@
-export { Pricing } from "./Pricing.tsx";
+export { Models } from "./Models.tsx";
 export type { Modalities, ModelPrice } from "./types.ts";
