+++ enter.pollinations.ai/src/client/components/news-faq/index.ts
@@ -1,0 +1,1 @@
+export { NewsFaq } from "./news-faq.tsx";
